workspace/chinese_tutor/requirements.txt
text
streamlit
chinese-converter
pypinyin
pandas
PyYAML
requests
chromadb
python-dotenv